Add AceToString function
AceToString provides the opposite functionality of ConvertToAce...it takes
a QualifiedAce and converts it into a string that's in the same format as
the strings ConvertToAce takes. Eventually, it can/will be used to show
the extra/missing access when -Detailed is provided to Test-Acl.

This is a first attempt at being able to convert an ACE into a string. In
theory, even non QualifiedAces should be able to work with this, but the
module should only require QualifiedAce support.

More Pester tests are required, especially for different enums and for
object ACEs, but there is one that can be easily extended included in this
commit.