Redo -RequiredAccess with temp SD method
We now build a temporary, blank SD for each ACE in -RequiredAccess, and
then call FindAce to find any ACEs that have overlap with the required
ACE, and we call RemoveAce with any matches ( calling RemoveAce with each
ACE in the ACL would work, but it's not as effecient as just calling it
with relevant ACEs that FindAce returns.

This has the awesome benefit of being able to tell the exact access that
is missing (which we can tell by any ACEs remaining in the temp SD), so
-Detailed can return a better picture of what's going on instead of just
adding the ACE we were testing for  (which is how it worked before).

When we were depending on AddAce and testing if the resulting SDDL matched
the starting SDDL, we would get false negatives when AddAccess/AddAudit
methods would unecessarily split an ACE when inheritance or propagation
flags didn't match.

FindAce was modified to search for any ACEs that fully contained the
access contained in the tested ACE, but that would cause problems if there
was no single ACE that contained all of the access being tested for, but
multiple ACEs could be combined. For example, imagine a SACL like this:
Audit S Everyone Delete O
Audit F Everyone FullControl

If you tested for 'Audit SF Delete O', there wouldn't be a single ACE that
granted that, so the FindAce method would  report that the tested ACE
didn't exist. The new method fixes this.