Add ART for ConvertToAce; change *Access param types
ConvertToAce can now take an -AccessRightType parameter. In order to use
this internally, the *Access parameters to Test-Acl had to be changed so
that the conversion to QualifiedAces takes place inside the function after
the common security descriptor has been created (since it has logic to take
Get-Acl objects and keep track of their AccessRightType). Right now, they
only take [string] objects, but they can eventually be expanded to take
[object] types instead if we want to go back to accepting other ACE objects.

TransformScriptAttribute was removed since it isn't needed anymore.

Modifying ConvertToAce also made some other errors jump out, e.g., some of
the tests tried to test a RegistrySecurity SD with invalid registryrights,
e.g. 'Read' instead of 'ReadKey'. These tests were found by the non-terminating
error messages, and they were fixed.

Finally, AceToString now throws a non-terminating error if an unknown accessmask
is used when an ART is provided. The invalid access mask will be included in the
string instead of a null section for the access mask (which was a bug).