Fix issue where ART was ignored on multiline string inputs
When ConvertToAce was called with the -AccessRightType and -InputObject
was a multiline string, the strings were split, and ConvertToAce was
called on each one without the original parameters being passed back to
the new call. For now, created a new hashtable to store the parameters,
but might eventually just edit the PSBoundParameters directly and remove
InputObject and splat them along (this would make a difference if more
parameters are added to ConvertToAce).

Also created a new test file for ConvertToAce. Will eventually split the
big test file up, and move all of its ConvertToAce tets to this new file.