Take object type into account for default inheritance flags
Default inheritance flags differ for RegistryKeys/ActiveDirectory objects
and file/folders. Files actually have no inheritance flags, but ConvertToAce
doesn't need to worry about that b/c the functions to add and remove rights
will remove flags if the SD is for a non-container object.

Registry keys and AD objects really only use the ContainerInherit flag in
practice (in both of those schemas, every object is a container), even though
ACEs CAN contain the child object inheritance flag. For this reason,
ConvertToAce will try to guess the default inheritance flags if no flags were
specified, and it will leave off the CO flag when needed.

Previously, O, CC, CO was the default for everything, and when access was
checked against that, CO was almost always seen as extra rights.

We'll have to see how many ACEs in the wild for keys and AD objects have CO
specified, and whether or not we need to try to fix that (even though in
those situations, you could consider that CO access on the object to be
extra that either needs to be removed, or the params to Test-Acl needs to
be updated to contain CO in the expected access)