Change RemoveAce behavior to remove Synchronize FSR
First, the handling of the Synchronize right needs to be refactored from
the groud up! At a minimum, the ToAccessMask function's -Action paramter
needs to take different strings.

This commit changed the -Action param from Remove to Add, which makes it
so that the Synchronize right will always be inserted into the AccessMask
for file/folder ACEs. This is to stop RemoveAce from leaving the Synchronize
right behind when testing for AllowedAccess.

The 'Remove' -Action still has a place for when it's time to convert an ACE
back into a string, but the whole process needs to be refactored to make more
sense...

This fixed some failing tests and didn't introduce any new failures.