Add PsPath to -Detailed return object
When the -Detailed switch is used now, the return object includes a text
representation of the path of the securable object. This comes from the
PsPath from $InputObject (or from the PsPath from Get-Acl when $InputObject
has to go through it first). The PsPath is run through Convert-Path to get
rid of the fully qualified PSProvider path.

This has a side effect of messing up registry paths so they can't be re-used
directly. This will be fixed up when this functionality makes its way into
the PowerShellAccessControl module.

The benefit of this is that you can now do recursive lookups of folders,
registry keys, and AD objects so you can find objects that don't meet the
requirements you need them to.