Add band-aid for Audit ACE detection in Test-Acl
When the -*Access parameters were changed to take strings (instead of
QualifiedAce objects that had a transform attribute), that meant that
the initial test for Audit ACEs (which is needed so we know if Get-Acl
needs to have the -Audit switch provided) broke. It checked ACE objects
for an audit AceQualifier. To fix it, we could move this test to after
the *Access params are converted, but then we have a chicken/egg problem
since those conversions happen after the security descriptor is created
so that we can use it to tell what type of securable object we're
dealing with.

We could do a conversion that we throw away, but then we'd do the conversion
twice, and this function is already WAY too slow.

For now, we're just searching for the 'Audit' string at the beginning of a
line. I don't like it, but it'll do for now.