Test to make sure Get-Acl retuned something
If -Audit was used, but the caller didn't have the ability to use
SeSecurityPrivilege right, then this would just be $null and script
execution would continue. (Problem was noticed when Test-Acl was called
by SYSTEM account, and Get-Acl didn't seem to throw an error. I think
Get-Acl might not throw errors for AD objects?)