Fix failing tests
This update fixes tests that were either just too noisy (two of them
just had errors captured in a variable, but they were expected to write
errors, so that's fine) or, in one case, a parameter to a constructor
was invalid. Not sure why these weren't fixed months ago when I was last
working on this module, but they're good now.