Make AD ACEs work properly with -AllowedAccess
The way -AllowedAccess works is that a copy of the SD has each ACE
specified in the -AllowedAccess parameter removed from it. Well, .NET
methods for removing access don't remove access in certain scenarios. I
don't think the failure to remove the access is a bug, though, since it
appears the developers just don't allow extra access to be removed if
there is a bigger, over-arching ACE that allows the same access.

What this update does is that NewCommonSecurityDescriptor (the helper
function that's responsible for copying the original SD) creates up to
4 separate SDs for AD ACEs. Each combination of [ObjectAceFlags] is used
as a key to a dictionary, and as each ACE that is to be added is
encountered, its [ObjectAceFlags] are used as a key to the dictionary.

Then, when RemoveAce is called, it is simply called against each of the
SDs. There was surprising little change needed (so far) inside of
Test-Acl. The biggest issue was that the hidden properties that were
being tucked away on the SD required a change so that they now look at
the first item in the potential $SD array (non-AD objects should only
have $SD be a single value object, but the [0] notation will still work
in that situation).

I was also expecting to have to re-combine the SDs into one before
showing the extra access, but that doesn't seem to be necessary. There
are probably some situations where that would be beneficial, though, so
I'll try to come up with more test cases to try to see if that step is
needed.